requirements.txt
watchdog>=4.0,<5.0
requests>=2.31,<3.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
